--- types/plugin/setUtcOffset.d.ts
@@ -1,4 +1,0 @@
-import { PluginFunc } from 'dayjs'
-
-declare const plugin: PluginFunc
-export = plugin
